[INTERNAL] Creates a new `Utility` module defining miscellaneous logic
