+++ archey/api.py
@@ -4,10 +4,11 @@ from datetime import datetime
 
 import json
 
-from typing import Dict, List  # pylint: disable=unused-import
+from typing import List
 
 from archey._version import __version__
 from archey.entry import Entry
+from archey.utility import Utility
 
 
 class API:
@@ -31,7 +32,7 @@ class API:
                 entry.name: entry.value for entry in self.entries
             },
             'meta': {
-                'version': self._version_to_semver_segments(__version__),
+                'version': Utility.version_to_semver_segments(__version__),
                 'date': datetime.now().isoformat(),
                 'count': len(self.entries)
             }
@@ -40,14 +41,4 @@ class API:
         return json.dumps(
             document,
             indent=((indent * 2) or None)
-        )
-
-    @staticmethod
-    def _version_to_semver_segments(version: str) -> tuple:
-        """Transforms string `version` to a tuple containing SemVer segments"""
-        return tuple(
-            map(
-                int,
-                version.lstrip('v').partition('-')[0].split('.')
-            )
         )
