+++ archey/test/test_archey_version.py
@@ -5,7 +5,7 @@ import unittest
 from archey._version import __version__
 
 
-class TestVersionUtil(unittest.TestCase):
+class TestVersion(unittest.TestCase):
     """Brain-dead test to verify the version string handling in Archey"""
     def test_load_version(self):
         """Test Archey's `__version__` internal attribute is correctly loadable"""
